refactor: replace password and email schema in shared folder
